Modify AutoLayout engine to successfully layout shape lists that are not fully connected. Also tweaked case for entering new binary FactTypes to make sure they flow to the right more often. refs #173
Modify the FactEditor to enable adding object types (any ObjectType chain with one or more ObjectTypes and no reading text will result in ObjectTypes only). Also enabled ability to select ObjectifiedFactTypeNameShape to get the ObjectType name shown in the FactEditor instead of treating it the same as the associated FactTypeShape. #331

SVN Revision: 1296