Eliminate existing intra-fact type splits from role paths on load. An intra-fact type split is a continuation of a fact type entered in a previous section of the role path which is then split to create compound conditions. The replacement is a newly rooted path with a correlation to a pathed role in the initial entry section of the path. This checkin eliminates the intra-fact type splits on load, but does not make them illegal internally (pending).
This is a more versatile construct because it allows conditions to be applied to any in-scope path nodes, not just the roles in the immediate fact type. This construct was previously supported, and is now the only way to do this.

This is an internal deprecation and will not have a visible effect on NORMA users.

SVN Revision: 1513