Enhance element grouping capabilities
Element grouping enhancements (and corresponding enhancements to the
survey capabilities used to populate the ORM model browser) are needed
to support the following capabilities:

 * An element should be able to switch expanded groups by changing a
survey answer. Previously, changing an answer for a question used
for grouping or sorting would throw.
 * A question with grouping display should be able to place elements
with 'not applicable' answers before the grouping headers for the
answered questions.
 * Group types could support either automatic or explicit population,
but there was no way for one group to provide both capabilities at
once. Note that this is an interface change, so existing extensions
need to be modified and recompiled.
 * There was no way for a group type to add properties to a grouped
element (when referenced in the group). Added the
IAnswerIndirectSurveyQuestion interface to support redirecting
answers to a question on a group to a corresponding grouping type.
This enables group types to add custom sorting and expansions inside
a group element expansion.
 * Enable a grouping type to add custom menu items to a grouped item.