Adjust RolePathOwner copy behavior
The distinction between an 'owned' LeadRolePath and a 'Shared' lead
role path is minimal. We have to pick one owner so the role path can
be serialized in exactly on place on disk. However, it doesn't really
matter which one we pick or if we switch owners.

The change affects copy and merge semantics for path owners (fact type,
subtype, and subquery derivation rules and join paths). A path can now
be matched to both directly owned and shared paths, and the path can
change ownership when it is dropped into a target model.

The reason for this is to eliminate the parent the owner of the role
path owner as a required dependency. For example, if fact type and A
(the owner) and B (sharing A's path) existing in a model and B is
dragged into a new model without A, then B becomes the designated owner
of the shared role path in the new model, and A is no longer considered
as part of the copy closure for B. Removing this closure dependency
reduces the size of the copy closure for path-sharing RolePathOwner
instances.

If an owner is not included as part of the copy closure, then an owner
is automatically chosen from among the shared owners. Similarly, the
latest merge can change the owner, with the old owner being thrown
back in the 'shared' pool.