Add inherent mandatory constraint to core model
An inherent mandatory constraint is defined as a mandatory constraint
that would be implicitly added to the model if it were not explicitly
defined.

This will be used in future revision to reduce the difference in
implementation mapping between implicit constraints and constraints
that were explicitly added by a user for clarity. This user preference
should not have an effect on generated code as it currently does.