Colorize with domain models ordered by dependency (#7)
Overload the IFrameworkServices.GetTypedDomainModelProviders method to
retrieve models ordered by dependency instead of by the (default) name.
This applies a topological sorting algorithm based on model dependency
attributes on the domain model types.

Add the SoftExtensionDependencyAttribute to allow a model to depend on
a target model when the target model is loaded--without forcing the
target model to load. This attribute is recognized by the dependency
ordering in GetTypedDomainModelProviders. Extensions are identified
by their model id.

Retrieve all uses of the IDynamicShapeColorProvider service with
dependency ordering, with the first model to return a color winning.
This allows multiple extensions to provide colorization with
predictable results based on soft dependencies between the extensions.