Temporarily do not serialize inherent mandatory
This is a preemptive fix to avoid a core file format change because
of the inherent mandatory feature. We will turn this back on before
integrating into the master branch. In the meantime, inherent
mandatory constraints are reevaluated on load, and this will still
load any previously saved files with the inherent mandatory link
saved.