VSIX support multiple verbalization directories
VS2017 (VSIX) installation does not support writing to directories
in another extension. This means that we cannot support verbalization
snippets in a single location. Instead, each registered VSIX extension
with snippets must register its own verbalization directory.

VerbalizationDir registration has been move the registry tree under
DesignerSettings with the ConvertersDir and SettingsFile attributes.

The VerbalizationManager.LoadFromRegistry method is not currently
available in VS2017 and will be redesigned in the future.