Verbalize smart quotes as character references
The left and right smart quote characters in verbalization snippets
are rendering correctly in Chrome verbalization reports but not in
Edge or IE. Use the decimal character references 8220 and 8221 instead
of the quote characters.

The most common construct that used smart quotes is the
ObjectifiesFactTypeSnippet, which displays when an objectifying object
type is selected.

I do not know why this worked in the verbalization browser, which is
rendered with a Microsoft browsers.