Use identifying role instance for implied fact (#14)
Fixes #13

When a fact type is part of a preferred identifier the population of
the fact type is implied by the population of the identified entity
instance. For a single-role identification scheme the object instance
and its (only) role instance display the same. However, for a compound
identifier we need to display the role instance on the identifying
role, while the identified role keeps the full instance display.

The fix extends the FactTypeInstanceImplication helper struct to
indicate if an identifying role in a compound identifier has been
selected. If this is set and the current column corresponds to the
identifying role then display the role instance instead of the
entity instance.