Do not wrap or clip Relational View table names
Fixes #15

Modify the table name text field rendering options to never clip, wrap
or display ellipsis. This could be triggered by small discrepancies in
measuring text that cause overflow pixels that are not visible to the
end user but could still overflow the official boundary and trigger a
text clipping response.