Subtypes in exclusion/mandatory contradictions
Fixes #19

An exclusion/mandatory contradiction should occur in a single-column
exclusion constraint if, given the role player on the mandatory
constraint, there is at least one other constrained role that
hosts the same role player or a subtype of the mandatory role player.

Similarly for equality constraints, if all constrained roles are also
mandatory then the equality constraint is only implied if all the
constraint roles host the same role player. Also modified this
implication validation to apply in alethic cases only.