Context Window and Generator Settings scaling
Fixes #21

The ORM Context Window creates the controls at the top of the window
with custom code, not with designer-generated files. I could not get
AutoScaling to trigger automatically with this approach. However, I
can get the value for the current font size and add custom scaling.
The Generations label and spinner controls now have enough vertical
space and no longer overlap.

The ORM Generator Settings control was also not scaling properly. This
was fixed by setting the AutoScaleMode to Font instead of Inherit.