Side-by-side build cleanup
The VsSdkLocator task was failing off of my machine because the
supporting build assembly was not found. It was only found on my
machine because I inadvertently put it in the GAC before the I finished
with the side-by-side installation work, which does not use the GAC.

The adjustments from AssemblyName to AssemblyFile in UsingTask
declarations in .targets files needs to be applied to VS2017 in
addition to VS2019.

The designer extensions and generator sections were still being
written to the registry in a side-by-side install. Modify install.bat
files to skip this write for side-by-side environments, where all of
the registry keys are written by the matching .pkgdef file during
the .vsix build.