New items appear outside nested neutral branch
Items were appearing both in expanded branches and at the end of
nested neutral branches. This was caused by not verifying that a
section had actual data in it before relying on the End count. If
the subbranch is empty, then Start>End. Subbranches in this state
do not update the Start/End values, so adding enough items will
eventually make End higher than the total node count, resulting
in items being added past the end of the list when relying solely
on the End count.