Rework nested grouping element insertion
The previous 'contextItemAdded' approach for managing inserted items
with nested groupers was not sufficient. This needs to be handled one
nesting level at a time, not with a status from multiple levels up.

There are two primary issues handled:
-When an item is added on the edge of a group section there is no way
to tell based solely on the index if the item was added to the section
above or below the insertion point. The answer depends on whether the
current item has a matching answer and whether the item was already
inserted previous in a previous section.
-When groups are nested, the answer to the inner group is irrelevant
unless the outer group has a neutral or matching answer. This could
result in an item being inserted in multiple sections.