git and user settings cleanup
Fixes #28
Modifies #27

Add .gitattributes to avoid adding any additional files under safecrlf
settings. At some point when there are fewer active branches we'll
normalize the random line ending and whitespace settings across the
project. The current mess was effectively caused by this evil setting
modifying line endings on checkin. The expected project settings are
CRLF line endings, leading tabs, and no trailing whitespace.

Added stock .gitignore with a few additional NORMA additions so there
is less 'untracked files' noise on git status and checkin.

Added .csproj.user files to all projects that need to directly launch
the VS experimental instance with F5. This was previously done by
copying the ORMModel\ORMModel.csproj.user file to the correct
directory, but this is not obvious to new developers. Updated the
readme to eliminate the copy instructions, and added the .user files
to .gitignore (these files are macro-based and should not be touched).