Extension properties exception closing model
The model tear down process can leave the VS properties window with
references to elements from the closed (or closing) model. Getting
properties from a zombied object leads to random exceptions.

In this case, the code to get extension properties off a standard
ORMModelElement is not verifying the object is in a valid state
to retrieve additional properties.