Cross model drag silent failure
Utility.GetTypedDomainModels returning null instead of an empty
enumerable if no extension implements ICopyClosureFilterProvider.

This interface was added to enable loaded extensions to filter
cross model drag-drop operations. This interface is not implemented
by the core (default loaded) models. This appears to be the first
such interface not implemented by the core models that hits this
code path. Make sure we return an empty enumerable instead of null
in this case.