Context change notification for survey elements
Enable a survey node that implements ISurveyNodeContext to notify that
the SurveyNodeContext may have changed. If the context has changed
then the old element is deleted (including any expansions) and the
element is attached to the new context.

This was needed because events for delete-then-add scenarios can
arrive in an undefined order. If the change to trigger an add event
arrives before the corresponding delete change then an attempt will
be made to add the item in two places. Conversely, if the delete
event triggers first then there is no way to track that the items
are actually the same and maintain a connection such as selection
state across the added and deleted items. (Selection tracking is
not in this commit.) By conditionally adding/deleting the item
in a single call it becomes safe to call this multiple times without
any danger that the operations will be done in the incorrect order.

INotifySurveyElementChanged.ElementContextChanged implements this
notification.

Also enabled EditorUtility.PropertyDescriptorAs to recognize
descriptors returned by EditorUtility.ModifyPropertyDescriptorDisplay.