Empty grouped neutral branch ignores first add
If a list is completely empty and has a neutral branch (so both
headers and items are displayed) then the case to recreate a null
neutral branch and add the first item to it was not checked correctly.

This occurred only after a delete sequence removed an empty neutral
branch (leaving it null) and no header groups were currently visible.