Enable soft dependencies before another model
NORMA load order is based on dependencies between models. We recently
added a 'soft dependency', which allowed a topological sort of
unrelated models without forcing the dependent model to load. However,
this always placed the soft dependency after the referenced model. We
extend this to place the model with the SoftExtensionDependency
attribute to also sort before the referenced model.

Also enabled the AlsoLoadDomainModel attribute. This is assumed to be
a non-serialized helper class with no independently identifiable URI,
where the URI generally used to load a normal model. This enables a
model to use a soft-dependency to place code both before and after a
referenced model by registry a primary and an 'AlsoLoad' model,
thus allowing model-implemented services to trigger a call both before
and after a referenced model in a topological ordering over models.

An example use is negotiating the 'winning' color when multiple
services provide colorization to diagram shapes.