VS2019/Win10 diagrams disappear adding extensions
fixes #30

Diagrams are all blank (except for the Diagram Spy) after changing
extensions on newer versions of Windows 10 (after 1703). This comes
from changes to Dpi handling when parent/child windows are attached
using the SetParent API, which fails in some cases after this change.

The fix recreates the DiagramClientView window handle. The better fix,
which Microsoft needs to do, is to not create this window so early in
the first place. There is no reason this needs to have a handle
created before the parent window is created.