Allow non-shape drops other than ElementGrouping
The ability to drop a group between stores was hard coded into the
diagram drop targets. Make this behavior available to extension
elements as well by implementing the IShapeFreeDataObjectProvider
interface on the corresponding extension domain model. This is used
to return an array of types (derived from ModelElement) that need to
support this drop behavior.