Keep Model Browser selection on context change
The ElementContextChanged event deletes and adds the element
internally. If the item was selected this will naturally remove that
selection, so it needs to be restored. However, the tree and branches
have no knowledge of selection, which is completely maintained by the
tree control.

Add an ElementLocationChanged event to the survey tree that will be
triggered when an ElementContextChanged notification actually moves
an element. This is exposed through the branches so it can be easily
accessed through the VirtualTreeControl.Tree.Root branch. The model
browser then attaches and listens to this callback during event
playback to restore the selection when all other changes are complete.