Less aggressive model browser selection tracking
In response to the ElementLocationChanged event the model browser
should reselect the element if it was previous selected. However,
it should not reactivate the window or change focus to the window,
which also happens with the NavigateTo command. Use the
VirtualTreeControl.SelectObject call directly instead of the
full NavigateTo command.