Expansion display for element context change
INotifySurveyElementChanged.ElementContextChanged can add/remove the
first/last item from a parent list. If the list has already been
expanded then the state is easily tracked. However, if no expansion
has been made then we need to tell the new parent to redraw its
expansion state if the new parent is visible in an expanded list.

To reliably move items we need to know the old parent node. This is
not available from the current survey objects, so we had to add a
parameter to ElementContextChanged to pass in this data.