Resolve dependent assembly types with extensions
Like extensions, dependent assemblies may also load from a directory
which is not on the assembly probing path for Visual Studio. Although
the assemblies are loading correctly, a Type.GetType call for a type
in the assembly will not always be found, depending on how the
assembly was loaded in the AppDomain.

This is causing the RectangleD type (in the SDK Diagrams assembly) to
load correctly, but the associated RectangleDConverer type converter
to not load from the same assembly. This means that any RectangleD
property (like AbsoluteBounds, which is on every diagram shape) cannot
be deserialized from XML.