Assembly location in property load failure assert
Failure to deserialize a property is generally due to the type
converter failing to load or loading from the wrong assembly. If the
wrong assembly loads it might actually have the correct assembly name
just come from the wrong location, so expand the assert to include
assembly file location information as well as assembly name.