Enable VS2019 Transaction Viewer filters
VS2019 added additional screen reader support for the VirtualTreeGrid.
However, they neglected to respect the 'DelayRedraw' setting on the
tree in the implementation, which blocks the UI from redrawing until
redraw is reenabled. This is used during a filter operation in the
transaction viewer dialog and means that out-of-bounds requests are
made by the screen reader while the tree is being reconfigured for
the filter operation.

This change tracks this condition and returns empty data for all
drawing requests (GetText, GetTipText, IsExpandable, GetDisplayData)
during the redraw operation.