Enable dynamic signing file for official builds
The ORMPackage.snk file used to sign NORMA assemblies is included in
the repository, which means that anyone can build the assemblies and
assemblies signed with signing file cannot be trusted to come from a
reliable source.

This change enables a pluggable signing file and introduces an
official version of tool, where the PublicKeyToken for the assemblies
is known but the corresponding signing file is not public.

Extension developers can include the NORMACommon.targets file the same
as the built-in NORMA projects (see changes in the CustomProperties
project as an example) and reference the official assemblies by
setting the NORMAOfficial=1 environment variable before building their
projects. This assumes that the setup version of NORMA has been
installed in the Visual Studio Experimental instance (launched with
"devenv /rootsuffix exp") and that the official NORMA assemblies,
which are now unbuildable because the official signing key is not
public, have been back-copied into the source tree with the new
NORMAAssembliesFromSetup.bat file added for this purpose.

The VSIXOnly project used for VS2019 and VS2017 automatically
incorporates the dynamic PublicKeyToken into the installers.

Setup will be modified for earlier versions in a later commit.