Change to git-based file versioning
The quarter-and-day based versioning system recently rolled over to
100 (for example Christmas 2020 was revision 10091). This five-digit
revision number was no longer compatible with the WIX-based setup
identifier generation thereby blocking setup from being built for
VS2015 and earlier versions of the product.

A ground-up rewrite of the VersionGenerator utility was done to
support this versioning approach. There are 4 enhancements:

1) The GitCommand and GitCommandArgs options have been added to
specify an external process to invoke to get the git-based file
version. This currently calls 'NORMAGitVer full .', which outputs
the expected version information.

2) VersionGenerator.exe no longer hard-codes specific files to
generate. The output files are now part of the configuration file in
the form of templates with replacement fields. A template element has
a format string as the contents of the element. The attributes on
the template element are output (the file name to write) and numeric
values for each of the available replacement fields, indicating the
replacement index to use in the format string. Available fields are
major, minor, build, revision, hash, yearMonth, buildType, and
warning. The Version.bat, Version.wxi and Version.cs files are now
templated in the config file instead of hard-wired into the utility.

3) Since git versioning can change without touching the day or
config files there was a need of a separate file to indicate the
state when the files were last generated. This is set with the
templateSettings@last attribute. This file is not checked in and
has been added to .gitignore.

4) VersionGenerator.exe can optionally take an argument indicating
the file path to an alternate configuration file. This allows
extension developers to use this utility to version assemblies using
the same approach taken here and/or to reference the NORMA version
information. The current directory is also set to the location of the
alternate configuration file.

With the git-based versioning, the MatchBuild = true setting for the
NORMAExtensionCompatibility attribute will be less useful because the
NORMA core and extension versions will be based on different git
histories. This means the NORMA build number will need to be
dynamically extracted from the NORMA repository to enable extension
compatibility checks. The VersionGenerator utility config-based
templates and custom configuration files enable this scenario.