Exclusive-or constraint link saved in wrong place
The ExclusiveOrExclusionConstraint element in a MandatoryConstraint
is saving after the ImpliedByObjectType and InherentForObjectType
elements. This causes a schema validation error on load because the
schema file specifies a different order.

For new files it doesn't matter whether we change the save code or the
schema, they just need to match. We'll update the schema file so that
previously saved files with this error will load without schema
validation errors.