Generate multiple files with one output format (#35)
Add the notion of 'Generator Target' to enable generation of multiple
files of the same type based on inputs from the current model. This is
a major enhancement from the prior one-file-per-output-format approach
and is designed for advanced model partitioning where a single file of
a given type cannot accurately represent the output. Doing this
successfully requires dynamic input from the model extensions, which
means that the ORM model must be loaded for code generation if a
generator signals support for generator targets.

The generator registers the GeneratorTargetTypes value (REG_MULTI_SZ)
indicating the target types that are needed to generate multiple files.
An extension domain model implements the IGeneratorTargetProvider to
provide GeneratorTarget (TargetType, TargetName, TargetId) elements to
the generation layer. Here the name is a file part and the id is passed
in a TargetType parameter to the transform. TargetTypes should not
overlap file format names as these are also passed to the transform.
Note that this means that running the custom tool can actually change
the output files whereas previously the set of output files was not
dependent in any way on the model contents.

In addition to the GeneratorTargetTypes registry values a generator
can also add GeneratorTargetInstruction_TYPE for each of the registered
target types. This allows generator-specific instructions to be sent
to the generator transforms. If there is no model-provided data
specifying the generator target we still generate a single file of that
type by sending modified parameters to the generator transform. The
_GeneratorPlaceholders parameter is added with a space-delimited list
of each of the target type names that do not have data provided by the
model. In this case if an instruction is registered it will be passed
in the corresponding target type parameter, which would otherwise be
empty. This mechanism allows a generator to either produce content for
the full model or simply provide the end user with instructions on how
to get content in the generated file(s).

There were some supporting framework changes needed to reasonably
enable this feature:

Enable an extension to be marked as non-generative. A non-generative
extension is not used for code generation. These are extensions used
for shape and model browser display and edits to other models. If an
extension does not serialize elements it should always be marked as
NonGenerative=1 in the extension registration. This is done for faster
code generation in the 'generator target' case, which requires a query
of the model itself to determine the files to generate. Given that
shape loading is the majority of the load time in most files, this
significantly improves performance when generating files that are not
already open in the ORM designer.

Replace CustomSerializedXmlNamespaces with new CustomSerializedXmlSchema
attribute. This adds the schema information needed to load the a
schema from an assembly resource without loading the domain model.
This is needed for non-generative loading because we still need the
xml to validate, even though we ignore these elements during load. This
change can be made by running the updated SerializationExtensions
transform, which will produce the correct attribute. The old attribute
is currently deprecated and will be removed in due time.

This change will not compile in VS2005, so this marks the end of the
road for VS2005, which has been officially unsupported by Microsoft
for almost 5 years. BuildItem.MetaDataNames is not available on
VS2005 and there is no reasonable workaround. I'll keep VS2008
builds alive for now.