Abstraction gateway leaking unidentified elements
The ORMElementGateway component is designed to keep incomplete ORM
elements (namely those with modeling errors and missing identification
schemes) from mapping through the abstraction layer to the relational
layer. The code was missing a case where incomplete indirectly
identified role players of an objectification were not being flagged
as incompletely, which incorrectly leaked the objectification through
the gateway for processing. The main processing could not complete
without the full identification schemes for these leaked elements.

This also fixed a case where an objectification had its primary
identifier on a supertype. In this case, the PreferredIdentifier
relationship for the object type is false, so we have to fall back
on ResolvedPreferredIdentifier, which gives back the identifying
constraint on the supertype.