Model loader aborts stripping all extensions
The model loader should survive loading an ORM file that uses
extensions that are not installed on the current users machine. The
correct behavior is to strip the extensions, load successfully, then
notify the user so they can avoid save a file with reduced data.

This worked fine unless all of the non-standard extensions failed to
load, which caused the loader to crash. Add additional null checking
to handle this edge case.

Also finished removing support for the CustomSerializedXmlNamespaces
attribute, which was obsolete and replaced by the
CustomSerializedXmlSchema attribute. This completes the support for
marking extensions as non-generative, producing faster code generation
times in cases where the file needs to be loaded to generate.

The GeneratorTarget support (which enables multiple outputs of the
same file format) had a couple of small errors that prevented the
generator targets from being passed between the designer and code
generator threads.