Enable PerUser installation
The only reason NORMA requires an AllUser (admin) install is to
support schema file installation in Visual Studio. However, the
ability to edit XML files (including the .ORM file as XML) in the
Visual Studio XML editor is a rarely used feature for most NORMA
users. Enable a PerUser build of the installer that does not
require admin rights, at the cost of losing automatic schema support.

The normal install and per-user install are not comatible. The user
must pick one. For now the vsix identifier has not changed, so only
the normal build will be in Visual Studio marketplace. I would have
preferred an 'incompatible with' setting, but this does not appear
to be available in the vsixmanifest file format.

Note that the NORMACatalog.xml file has been modified, so opening
$Schemas\NORMACatalog.xml from the per-user extension installation
directory will enable all of the other NORMA schemas in the current
Visual Studio session.