Enable custom properties on an ORM Diagram
Satisfy a request to attach a custom property to a diagram in addition
to the ORM model element.

Adding diagram support to the custom properties is easily done once
the diagram supports extension elements, which is is how the custom
properties are stored. Adding extension element support to something
other than ORMModelElement requires new relationships (on ORMDiagram
and ORMBaseShape for completeness) to support storing extension
elements with these objects. It also requires a new mechanism for
informing the system about these new relationships. This is done with
the new IORMExtendableElementService, which is available from a
corresponding property on IORMToolServices.

The diagram element also did not have any native verbalization support.
Like other shapes it simply defers to the represented model element.
This will now try to verbalize the diagram element (which now just
consists of the custom children) before verbalizing the orm model. The
result is that diagram custom properties will verbalize first,
followed immediately by the traditional model verbalization.

No attempt is made to block adding the same custom property in both
the 'ORM Model' and 'ORM Diagram' categories. However, this is not
recommended as it will give you two identical properties with a
diagram selection. One will be for the diagram and the other for the
model, but you won't be able to tell which one is which.