Add Error List Window is an ORM Tool Window
Add the Visual Studio Error List Window to the ORM Tool Windows
menu.

This and the Properties Window, which is already at the bottom of the
menu, are the two Visual Studio tool windows that are frequently used
while editing ORM models. This increases discoverability of this
important tool window.