Reference mode naming option changes on reload
If an ReferenceModeNaming instance is associated with an object type
and has non-default property values (which forces serialization), then
the PrimaryIdentifierNamingChoice attribute will switch from
ModelDefault to ValueTypeName on reload.

The default value in the schema file does not match the default value
in the serialization. Since the attribute does not write, the
validating schema changes it to the default in the schema file during
deserialization. Fix the incorrect default in the schema file.