Block absorption of an empty relational table
NORMA does not add extra columns for 'isSUBTYPE' concepts, so if an
asserted subtype is absorbed and plays no additional roles (beyond
the subtyping role) then it simply disappears on assertion.

The optimal solution is to automatically add columns to explicitly
assert the subtype. However, this is far from trivial. First, the data
structures do not exist in either of the mapping layers. Second, the
generated relational column would also need to apply to subtypes with
no mandatory roles.

Given that a subtype that plays no roles is generally considered an
anti-pattern and that this can be easily worked by using a unary
fact type instead of a subtype this case is not worth additional
resources. This mostly happens in small demo models, not fully fleshed
out models with actual roles on the subtypes.

This increased the core algorithm version, which means that files will
automatically recreate the relational mapping during reload.