Extension build fails with official release
Official releases are built with a private signing file which is not
provided to external extension developers. To build against the
official version it must be installed with setup into the Experimental
VS hive (used for developing VS with VS). The installed files are then
back-copied into the NORMA build output directories where they are
referenced by the extension projects.

The batch file to do the back copy had a false negative verifying a
current NORMA installation. This blocked the build process.