AlsoLoadDomainModel load fails silently
Fix a defect introduced with recent changes to support non-generative
load options with CustomSerializedXmlSchemaAttribute replacing
CustomSerializedXmlNamespacesAttribute. The change accidentally
required each extension to provide a serialization namespace and
silently disabled the extension otherwise.

The fix reverts to the prior behavior, which requires the extension
namespace to match a serialized namespace if the extension serializes
elements.

This did not apply to AutoLoad extensions, so only affected secondary
extensions loaded with the AlsoLoadDomainModel attribute.