Move common name generator concepts to core
Centralize name generation helpers for customizing reference mode
names, combining name parts, modifying name parts for case and common
phrases, and producing increasingly specific names as part of a unique
name generation process.

These were previously implement inside the relational model bridge, but
there were so many concepts common to any name generator that it made
sense to refactor and share the code.

ReferenceModeNaming is now a set of generic classes used for gathering
data that can be easily used by extensions by providing class and link
information for the global and targeted customization instances.

Also added the notion of refined settings to object type and recognized
phrase alias to enable multiple generated outputs of the same type to
use different customizations.

This also fixes a relational issue where a reference mode naming
setting was lost on reload. This occurred because of a mismatch between
the default setting in code and the default value specified in the
schema file.