Non-generative secondary model generation failure
A non-generative secondary domain model (loaded with the
AlsoLoadDomainModel attribute) throws attempting to verify extensions
for a generative-only load.

Usually all known extension models are available. However, for a load
that ignores extensions marked as NonGenerative a secondary extension
may not be in the list of available extensions. This was causing the
VerifyExtensions method to fail, which aborted the code generation
process when a code generator required specific extensions.

Also fixed a save issue removing items from the generator settings
dialog.