Dynamic customtool generation list loop failure
An inner loop on the ORMCustomTool generator was not correctly
breaking to continue with the outer loop. This was an error in recent
changes to allow dynamic additions to the list during generation.

The likelihood of hitting this was relatively low with a small number
of generation targets, but increased dramatically as more were added.