Relational reference mode naming defaults ignored
This is related to the recent reference mode naming refactoring.

The ReferenceModeNaming.ResolveObjectTypeName needs some element to be
passed as the first argument to be able to resolve the partition to
retrieve the global reference mode naming instance when a refined
instance as not defined. The relational code was passing null here,
which meant that the per-ObjectType ReferenceModeNaming settings were
resolved, but not the global DefaultReferenceModeNaming instance.

This is easily fixed passing the context ORMModel instance.